fix invalid UDP response packets
If the response is too big for an UDP packet, the TC flag is set to 1
and the packet is reduced in size to fit into the UDP packet.
With current bind-utils, this leads to this:

~> nslookup foo.bar.baz.biz 127.0.0.1
;; Got bad packet: unexpected end of input
512 bytes
9c c9 83 80 00 01 00 3d 00 0d 00 00 03 66 6f 6f .......=.....foo
03 62 61 72 03 62 61 7a 03 62 69 7a 00 00 01 00 .bar.baz.biz....
...
We see that the ANCOUNT is 0x3d and NSCOUNT is 0x0d, which makes
nslookup expect the actual packet to contain this number of responses,
but it doesn't, it is just truncated.

With tcpdump, this part of the answer is:
 0x0020: 0001 003d 000d              QDCOUNT=1 ANCOUNT=0x3d NSCOUNT=0xd
and the end of the 540 bytes packet is
 0x0200: 0100 0100 0002 3100 04xx xxxx xxc0 0c00
 0x0210: 0100 0100 0002 3100 04xx xxxx

When asking the "upstream" bind server, the answer is:
 0x0020: 0001 001d 0000              QDCOUNT=1 ANCOUNT=0x1d NSCOUNT=0
the 525 bytes packet ends with
 0x0200: 0100 0100 0002 5800 04xx xxxx xx
after exactly 29 (0x1d) A records.

Fix this by refusing to add records in add_rr() if the resulting size
would exceed the size given in the "udp" parameter. To signal that the
packet was truncated, set "udp" to zero and use that to set the TC flag
in the response packet.